101 fixing failed tests of test analysispy (#102)
* fix: incorrect slicing of Video when providing np.ndarray

* fixes various failed tests of analysis.py
